+++ Demos/AllOverIt/Patterns/PipelineAsyncDemo/Program.cs
@@ -3,11 +3,12 @@ using AllOverIt.Patterns.Pipeline.Extensions;
 using Microsoft.Extensions.DependencyInjection;
 using PipelineAsyncDemo.Steps;
 using System;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace PipelineAsyncDemo
 {
-    public delegate Task<string> IntegerPipelineProcessorAsync(int value);
+    public delegate Task<string> IntegerPipelineProcessorAsync(int value, CancellationToken cancellationToken);
 
     internal class Program
     {
@@ -27,19 +28,17 @@ namespace PipelineAsyncDemo
             var r1 = p1.Invoke();
 
             var p2 = PipelineBuilder
-                .PipeAsync(() => Task.FromResult(1))
-                .PipeAsync(v => Task.FromResult(v * 2.0d))
+                .PipeAsync(cancellationToken => Task.FromResult(1))
+                .PipeAsync((v, cancellationToken) => Task.FromResult(v * 2.0d))
                 .Pipe(v => (int)(v * 3))
                 .Pipe(v => v * 3)
                 .PipeAsync(new DoubleStepAsync())
                 .Pipe(v => v * 3)
-                .PipeAsync(v => Task.FromResult(v * 2.0d))
-                .PipeAsync(v => Task.FromResult(v.ToString()))
+                .PipeAsync((v, cancellationToken) => Task.FromResult(v * 2.0d))
+                .PipeAsync((v, cancellationToken) => Task.FromResult(v.ToString()))
                 .Build();
 
-            var r2 = await p2.Invoke();
-
-
+            var r2 = await p2.Invoke(CancellationToken.None);
 
             var services = new ServiceCollection();
 
@@ -83,7 +82,10 @@ namespace PipelineAsyncDemo
 
             var pipeline = serviceProvider.GetRequiredService<IntegerPipelineProcessorAsync>();
 
-            var result = await pipeline.Invoke(5);      // Will return 5 * 2 * 2 * 2 * 2 + 0.2 = 80.2
+            var cts = new CancellationTokenSource();
+            cts.Cancel();
+
+            var result = await pipeline.Invoke(5, cts.Token /*CancellationToken.None*/);      // Will return 5 * 2 * 2 * 2 * 2 + 0.2 = 80.2
 
             Console.WriteLine($"{result}");