+++ Source/AllOverIt/Patterns/Pipeline/PipelineBuilderAsync.cs
@@ -1,5 +1,6 @@
 ﻿using AllOverIt.Assertion;
 using System;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace AllOverIt.Patterns.Pipeline
@@ -7,14 +8,14 @@ namespace AllOverIt.Patterns.Pipeline
     // Begins a new asynchronous pipline that takes an initial input
     internal sealed class PipelineBuilderAsync<TIn, TOut> : IPipelineBuilderAsync<TIn, TOut>
     {
-        private readonly Func<TIn, Task<TOut>> _step;
+        private readonly Func<TIn, CancellationToken, Task<TOut>> _step;
 
-        public PipelineBuilderAsync(Func<TIn, Task<TOut>> step)
+        public PipelineBuilderAsync(Func<TIn, CancellationToken, Task<TOut>> step)
         {
             _step = step.WhenNotNull(nameof(step));
         }
 
-        public Func<TIn, Task<TOut>> Build()
+        public Func<TIn, CancellationToken, Task<TOut>> Build()
         {
             return _step;
         }
@@ -24,26 +25,26 @@ namespace AllOverIt.Patterns.Pipeline
     internal sealed class PipelineBuilderAsync<TIn, TPrevOut, TNextOut> : IPipelineBuilderAsync<TIn, TNextOut>
     {
         private readonly IPipelineBuilderAsync<TIn, TPrevOut> _prevStep;
-        private readonly Func<TPrevOut, Task<TNextOut>> _step;
+        private readonly Func<TPrevOut, CancellationToken, Task<TNextOut>> _step;
 
-        public PipelineBuilderAsync(IPipelineBuilderAsync<TIn, TPrevOut> prevStep, Func<TPrevOut, Task<TNextOut>> step)
+        public PipelineBuilderAsync(IPipelineBuilderAsync<TIn, TPrevOut> prevStep, Func<TPrevOut, CancellationToken, Task<TNextOut>> step)
         {
             _prevStep = prevStep.WhenNotNull(nameof(prevStep));
             _step = step.WhenNotNull(nameof(step));
         }
 
         // Create a func that invokes the previous func and uses its result as the input to the next func (step)
-        public Func<TIn, Task<TNextOut>> Build()
+        public Func<TIn, CancellationToken, Task<TNextOut>> Build()
         {
-            async Task<TNextOut> func(TIn input)
+            async Task<TNextOut> func(TIn input, CancellationToken cancellationToken)
             {
                 var prevOutput = await _prevStep
                     .Build()
-                    .Invoke(input)
+                    .Invoke(input, cancellationToken)
                     .ConfigureAwait(false);
 
                 return await _step
-                    .Invoke(prevOutput)
+                    .Invoke(prevOutput, cancellationToken)
                     .ConfigureAwait(false);
             }